+++ app/src/components/auth/index.tsx
@@ -1,29 +1,4 @@
-import React, { PropsWithChildren, ReactNode } from "react";
-
-import { useViewer } from "@phoenix/contexts/ViewerContext";
-
 export * from "./CreateAPIKeyDialog";
 export * from "./OneTimeAPIKeyDialog";
 export * from "./DeleteAPIKeyButton";
-
-type AuthGuardProps = {
-  fallback?: ReactNode;
-};
-export function IsAuthenticated(props: PropsWithChildren<AuthGuardProps>) {
-  const { fallback = null, children } = props;
-  const { viewer } = useViewer();
-  if (!viewer) {
-    return <>{fallback}</>;
-  }
-  return children;
-}
-
-export function IsAdmin(props: PropsWithChildren<AuthGuardProps>) {
-  const { fallback = null, children } = props;
-  const { viewer } = useViewer();
-  // If the viewer is not an admin, show the fallback
-  if (!viewer || viewer.role.name !== "ADMIN") {
-    return <>{fallback}</>;
-  }
-  return children;
-}
+export * from "./AuthGuard";
