+++ app/src/pages/profile/ProfilePage.tsx
@@ -1,28 +1,11 @@
-import React, { ReactNode, useCallback, useState } from "react";
-import { graphql, useMutation } from "react-relay";
+import React from "react";
 import { css } from "@emotion/react";
 
-import {
-  Button,
-  Card,
-  DialogContainer,
-  Flex,
-  Form,
-  Icon,
-  Icons,
-  TextField,
-} from "@arizeai/components";
+import { Card, Flex, Form, TextField } from "@arizeai/components";
 
-import {
-  APIKeyFormParams,
-  CreateAPIKeyDialog,
-  OneTimeAPIKeyDialog,
-} from "@phoenix/components/auth";
-import { useNotifyError } from "@phoenix/contexts";
 import { useViewer } from "@phoenix/contexts/ViewerContext";
 
-import { ProfilePageCreateUserAPIKeyMutation } from "./__generated__/ProfilePageCreateUserAPIKeyMutation.graphql";
-import { APIKeysTable } from "./APIKeysTable";
+import { ViewerAPIKeys } from "./ViewerAPIKeys";
 
 const profilePageCSS = css`
   overflow-y: auto;
@@ -39,51 +22,7 @@ const profilePageInnerCSS = css`
 `;
 
 export function ProfilePage() {
-  const [dialog, setDialog] = useState<ReactNode>(null);
-  const notifyError = useNotifyError();
   const { viewer } = useViewer();
-
-  const [commit, isCommitting] =
-    useMutation<ProfilePageCreateUserAPIKeyMutation>(graphql`
-      mutation ProfilePageCreateUserAPIKeyMutation(
-        $input: CreateUserApiKeyInput!
-      ) {
-        createUserApiKey(input: $input) {
-          jwt
-          apiKey {
-            id
-            user {
-              ...APIKeysTableFragment
-            }
-          }
-        }
-      }
-    `);
-
-  const onSubmit = useCallback(
-    (data: APIKeyFormParams) => {
-      commit({
-        variables: {
-          input: {
-            ...data,
-            expiresAt: data.expiresAt || null,
-          },
-        },
-        onCompleted: (response) => {
-          setDialog(
-            <OneTimeAPIKeyDialog jwt={response.createUserApiKey.jwt} />
-          );
-        },
-        onError: (error) => {
-          notifyError({
-            title: "Error creating API key",
-            message: error.message,
-          });
-        },
-      });
-    },
-    [commit, notifyError]
-  );
 
   if (!viewer) {
     return null;
@@ -109,38 +48,8 @@ export function ProfilePage() {
               />
             </Form>
           </Card>
-          <Card
-            title="API Keys"
-            variant="compact"
-            bodyStyle={{ padding: 0 }}
-            extra={
-              <Button
-                variant="default"
-                size="compact"
-                icon={<Icon svg={<Icons.PlusCircleOutline />} />}
-                onClick={() =>
-                  setDialog(
-                    <CreateAPIKeyDialog
-                      onSubmit={onSubmit}
-                      isCommitting={isCommitting}
-                    />
-                  )
-                }
-              >
-                New Key
-              </Button>
-            }
-          >
-            <APIKeysTable query={viewer} />
-          </Card>
+          <ViewerAPIKeys viewer={viewer} />
         </Flex>
-        <DialogContainer
-          onDismiss={() => {
-            setDialog(null);
-          }}
-        >
-          {dialog}
-        </DialogContainer>
       </div>
     </main>
   );
